Add and remove some oh-my-zsh plugins
Added:
- tmux - allows auto start and close of tmux when opening a new terminal
window, plus a few aliases and fixes. Replaces old tmux hacks.
- vi-mode - adds vi-mode so I can go into Normal instead of Insert mode
and use vim movement keys etc.
- history-substring-search - start typing a command and then press
up/down arrows to search for matches in history. A version of this
already worked with no plugins but vi-mode breaks this. This plugin
fixes it.

Removed:
- bower - I no longer work in any projects using bower
- encode64 - I never really used this
- last-working-dir - This always got stuck on a single directory rather
than actually opening a new tmux window in the last directory I went to.
- sudo - I rarely use this and it doesn't play nice with vi-mode anyway.
- vagrant - I no longer work in any projects using vagrant
- web-search - I never really used this