Rework for MacOS + switch to neovim
I now have a MacBook for work, so have adapted my dotfiles to work on
both MacOS (tested) and Ubuntu (not yet tested with these changes). Not
much needed to change to do this - mostly using brew instead of apt for
installing packages.

Setting up a clean machine also gave me the opportunity to make a few
other changes I've wanted to do for a while:
- Switched from Vim to NeoVim. Since there is so much crossover now
  between my vim config and other dotfiles, the nvim config is included
  directly in this repo and the separate vim-config repo will be
  abandoned.
- Prefer putting config under ~/.config where supported, rather than
  cluttering the root of my home directory with a bunch of separate
  files/directories for various packages.
- Install Node.js properly via NVM rather than apt.