Install dropbar.nvim
This plugin seems awesome and replaces nvim-treesitter-context with a
nicer UI that only takes up a single line, and more interactive
features.

Also getting rid of nvim-tree because I never use it and this dropbar
allows me to navigate the filesystem and view sibling files etc if I
want to.