Use different target for rust-analyzer cargo builds
This stops the LSP from clashing with docker (or anything else) that is
also watching for changed files and recompiling in rust projects.

Without this, one build process has to wait for the other. Also, if the
two are using different compiler versions, it takes longer as all the
crates need to be recompiled for any change in the project files. Plus,
when docker starts a build it deletes the proc-macros which causes
rust-analayzer to complain that they are missing.